+++ include/nodepp/url.h
@@ -48,14 +48,14 @@ struct url_t {
 namespace url {
 
 /*────────────────────────────────────────────────────────────────────────────*/
-    
+
     inline map_t<string_t,uint>& protocols() {
     static map_t<string_t,uint>  out ({
          { "https", 443 }, { "wss" , 443 },
          { "tls"  , 443 }, { "dtls", 443 },
          { "http" ,  80 }, { "ws"  ,  80 },
          { "tcp"  ,  80 }, { "udp" ,  80 },
-         { "ftp"  ,  21 }, { "ssh" ,  22 } 
+         { "ftp"  ,  21 }, { "ssh" ,  22 }
     }); return out; }
 
     inline bool is_valid( const string_t& URL ){
@@ -65,7 +65,7 @@ namespace url {
 
     /*.........................................................................*/
 
-    inline string_t normalize ( string_t msg ) { 
+    inline string_t normalize ( string_t msg ) {
         string_t res = msg; static regex_t reg( "%[a-z0-9]{2}", true );
         while( reg.test( res ) ){
             auto data = reg.match( res );
@@ -77,7 +77,7 @@ namespace url {
 
     /*.........................................................................*/
 
-    inline string_t unnormalize ( string_t msg ) { 
+    inline string_t unnormalize ( string_t msg ) {
         string_t res = msg; static regex_t reg( "[^a-z0-9%]", true );
         while( reg.test( res ) ){
             auto data = reg.match( res );
@@ -85,46 +85,46 @@ namespace url {
             res = regex::replace_all( res, data, "%"+hex );
         }   return res;
     }
-    
+
     /*─······································································─*/
 
-    inline string_t protocol( const string_t& URL ){ 
+    inline string_t protocol( const string_t& URL ){
         string_t null; static regex_t _a("^[^:]+");
-        if( !is_valid(URL) || !_a.test( URL ) ) 
+        if( !is_valid(URL) || !_a.test( URL ) )
           { return null; } null = _a.match( URL );
             return null;
     }
-    
+
     /*─······································································─*/
 
-    inline string_t auth( const string_t& URL ){ 
+    inline string_t auth( const string_t& URL ){
         string_t null; static regex_t _a("//\\w+:\\w+@");
-        if( !is_valid(URL) || !_a.test( URL ) ) 
+        if( !is_valid(URL) || !_a.test( URL ) )
           { return null; } null = _a.match( URL );
             return null.slice( 2, -1 );
     }
 
-    inline string_t user( const string_t& URL ){ string_t null; 
+    inline string_t user( const string_t& URL ){ string_t null;
         auto data = string::split( auth( URL ), ':' );
         if( data.size() != 2 ){ return null; } return data[0];
     }
 
-    inline string_t pass( const string_t& URL ){ string_t null; 
+    inline string_t pass( const string_t& URL ){ string_t null;
         auto data = string::split( auth( URL ), ':' );
         if( data.size() != 2 ){ return null; } return data[1];
     }
-    
+
     /*─······································································─*/
 
-    inline string_t hash( const string_t& URL ){ 
+    inline string_t hash( const string_t& URL ){
         string_t null; static regex_t _a("#[^?]*");
-        if( !is_valid(URL) || !_a.test( URL ) ) 
+        if( !is_valid(URL) || !_a.test( URL ) )
           { return null; } return _a.match( URL );
     }
 
-    inline string_t search( const string_t& URL ){ 
+    inline string_t search( const string_t& URL ){
         string_t null; static regex_t _a("\\?[^#]*");
-        if( !is_valid(URL) || !_a.test( URL ) ) 
+        if( !is_valid(URL) || !_a.test( URL ) )
           { return null; } return _a.match( URL );
     }
 
@@ -141,7 +141,7 @@ namespace url {
 	         return null.empty() ? "/" : null;
     }
 
-    inline string_t host( const string_t& URL ){ 
+    inline string_t host( const string_t& URL ){
         static regex_t _a("[/@][^/#?]+");
         if(!is_valid(URL) ){ return nullptr; }
             auto data = _a.match( URL ).slice(1);
@@ -150,15 +150,15 @@ namespace url {
         else return data;
     }
 
-    inline string_t hostname( const string_t& URL ){ 
+    inline string_t hostname( const string_t& URL ){
         string_t null = host(URL); static regex_t _a("[^:]+");
-        if( !is_valid(URL) || !_a.test( null ) ) 
+        if( !is_valid(URL) || !_a.test( null ) )
           { return null; } return _a.match( null );
     }
-    
+
     /*─······································································─*/
 
-    inline uint port( const string_t& URL ){ 
+    inline uint port( const string_t& URL ){
 
         string_t _prot = protocol( URL );
         auto     _list = protocols();
@@ -170,15 +170,15 @@ namespace url {
             return string::to_uint( _a.match( _host ).slice(1) );
         } elif( !_prot.empty() ) {
             if( _list.has(_prot) ){ return _list[_prot]; }
-        }   
-        
+        }
+
         return 8000;
     }
-    
+
     /*─······································································─*/
 
     inline query_t query( const string_t& URL ){ return query::parse( search(URL) ); }
-    
+
     /*─······································································─*/
 
     inline url_t parse( const string_t& URL ){ url_t data;
@@ -190,7 +190,7 @@ namespace url {
         data.origin   = origin( URL );
 	    data.query    = query( URL );
         data.pathname = path( URL );
-        data.port     = port( URL );
+        data.port     = (int)port( URL );
         data.host     = host( URL );
         data.hash     = hash( URL );
         data.user     = user( URL );
@@ -199,27 +199,27 @@ namespace url {
         data.href     = URL;
 
         data.path = data.pathname + data.search;
-        return data; 
+        return data;
     }
-    
+
     /*─······································································─*/
 
-    inline string_t format( const url_t& obj ){ string_t _url; 
+    inline string_t format( const url_t& obj ){ string_t _url;
 
         if( !obj.href.empty() ){
             _url += obj.href;
-        } elif( !obj.origin.empty() ){ 
-            _url += obj.origin; 
-        } else { 
+        } elif( !obj.origin.empty() ){
+            _url += obj.origin;
+        } else {
             _url += obj.protocol + "//";
             _url += obj.auth + "@";
             _url += obj.host;
         }
 
-        if( !obj.path.empty() ){ 
-            _url += obj.path; 
+        if( !obj.path.empty() ){
+            _url += obj.path;
         } else {
-            _url += obj.pathname; 
+            _url += obj.pathname;
         }
 
         if( !obj.search.empty() ){
